@@ -3,8 +3,9 @@ import os from 'os';
 import fs from 'fs';
 import axios from 'axios'
 import yaml from 'js-yaml';
+import React from 'react';
 import { exec } from 'child_process';
-
+import uniqby from 'lodash/uniqby';
 
 ////////////////////Global Variables/////////////////////////////
 export var isWin = /^win/.test(process.platform);
@@ -13,44 +14,113 @@ export var dir = isWin ? homedir + '\\AppData\\Roaming\\dtt\\' : homedir + '/.dt
 export var prefix = isWin ? dir + 'gdc-client.exe ' : dir + './gdc-client ';
 
 //////////////////////Download Functions/////////////////////////
-export var getDownloadPrefs = (downloadFolder) => {
+export var getDownloadPrefs = () => {
   var prefStr = '';
   var numClientCons = 0;
   var prefs = yaml.load(fs.readFileSync(dir + 'prefs.yml', 'utf8'));
-  for (var sectionKey in prefs.parameters) {
-    if (sectionKey !== 'uploadParams') {
-      var section = prefs.parameters[sectionKey]
-      for (var obj in section) {
-        if (obj === 'numClientCons') {
-          numClientCons = parseInt(section[obj]);
-        }
-        else if (section[obj] !== false) prefStr += section[obj]
-      }
-    }
-  }
-  if (numClientCons > 6) numClientCons = 6;
-  var downloadStr = ' -d ' + fixSpace(downloadFolder) + ' ';
-  var tokenStr = ' -t ' + dir + 'token.txt ';
-  var strList = [downloadStr, tokenStr, prefStr];
-  return [strList, numClientCons];
 
+  prefStr = Object.keys(prefs.parameters).reduce((str, sectionKey) => {
+    if (sectionKey !== 'uploadParams') {
+      var section = prefs.parameters[sectionKey];
+      return Object.keys(section).reduce((str, obj) => {
+        if (obj !== 'numClientCons' && section[obj] !== false) {
+          return str + section[obj];
+        }
+        return str;
+      }, str)
+    }
+    return str;
+  }, '')
+  console.log(prefStr);
+  if (numClientCons > 6) numClientCons = 6;
+  var tokenStr = fs.existsSync(dir + 'token.txt') ? ' -t ' + dir + 'token.txt ' : ' ';
+  var strList = [tokenStr, prefStr];
+  return strList;
 }
 
-export var requestDownloadStatuses = (isUUID, arg) => {
+export var checkValidManifest = (manifests) => {
+  var message = [];
+  var excludedFiles = []
+  var validUUID = /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/;
+  return Promise.all(manifests.map(manifest => {
+    return new Promise((res, reject) => {
+      fs.readFile(manifest, 'utf8', (err, content) => {
+        if (err) {
+          reject(err)
+          return;
+        }
+        if (content.split('\n').slice(1).length === 0) {
+          excludedFiles.push(manifest)
+          message.push(manifest + ' has an invalid format\n');
+        }//forEach below will not run if length is 0
+        content.split('\n').slice(1).forEach(x => {
+          var columns = x.split('\t');
+          if (!validUUID.test(columns[0]) || !parseInt(columns[3])) {//if first column is not uuid or fourth is not a number
+            excludedFiles.push(manifest)
+            message.push(manifest + ' has an invalid format\n');
+          }
+        });
+        res();
+      })
+    }).catch(() => {
+      excludedFiles.push(manifests);
+      message.push('error reading ' + manifest + '\n');
+    })
+  }))
+    .then(() => [uniqby(excludedFiles), uniqby(message)])
+}
 
-  if (isUUID) {
-    return Promise.all(arg.filter(id => /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/.test(id)).map(id => {
-      return axios.get('https://api.gdc.cancer.gov/v0/files/' + id + '?expand=metadata_files&fields=file_size')
-        .then(res => {
-          return { uuid: id, status: 'Not Started', time: '', size: formatBytes(res.data.data.file_size), speed: '' };
-        })
-    }));
-  }
-  else {
-    var fileContent = fs.readFileSync(arg, 'utf8');
-    var fileInfo = fileContent.split('\n').slice(1, Infinity).map(x => x.split('\t'));
-    return Promise.all(fileInfo.map(x => ({ uuid: x[0], time: '', status: 'Not Started', size: formatBytes(x[3]), speed: '' })));
-  }
+export var requestDownloadStatuses = (uuids, manifests, relFiles, anns) => {
+  var validUUID = /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/;
+  return Promise.all(uuids.filter(id => validUUID.test(id)).map(id => {
+    return axios.get('https://api.gdc.cancer.gov/v0/files/' + id + '?expand=metadata_files&fields=file_size,file_name,access')
+      .then(res => {
+        return {
+          uuid: id,
+          status: 'Not Started',
+          time: '', size: formatBytes(res.data.data.file_size),
+          speed: '',
+          rel: relFiles.toString(),
+          ann: anns.toString(),
+          name: res.data.data.file_name,
+          access: res.data.data.access,
+          indivDownload: false,
+        };
+      })
+  }))
+    .then(objs => {
+      var statusObjs = objs;
+      try {
+        return Promise.all(manifests.map(manifest => {
+          return new Promise((resolve) => {
+            fs.readFile(manifest, 'utf8', (err, content) => {
+              var fileInfo = content.split('\n').slice(1).map(x => x.split('\t'));
+              Promise.all(fileInfo.map(x => {
+                return axios.get('https://api.gdc.cancer.gov/v0/files/' + x[0] + '?expand=metadata_files&fields=file_name,access')
+                  .then(res => {
+                    return {
+                      uuid: x[0],
+                      time: '',
+                      status: 'Not Started',
+                      size: formatBytes(x[3]),
+                      speed: '',
+                      rel: relFiles.toString(),
+                      ann: anns.toString(),
+                      name: res.data.data.file_name,
+                      access: res.data.data.access,
+                      indivDownload: false,
+                    }
+                  })
+              }))
+                .then(rows => resolve(rows));
+            });
+          });
+        }))
+          .then((rows) => {
+            return rows.reduce((acc, row) => acc.concat(row), statusObjs);
+          });
+      } catch (e) { return statusObjs }
+    });
 }
 
 
@@ -64,7 +134,7 @@ export var getUploadPrefs = (uploadFolder) => {
       var section = prefs.parameters[sectionKey]
       for (var obj in section) {
         if (obj === 'numClientCons') {
-          numClientCons = parseInt(section[obj]);
+          numClientCons = parseInt(section[obj], 10);
         }
         else if (section[obj] !== false) prefStr += section[obj]
       }
@@ -94,12 +164,32 @@ export var requestUploadStatus = (isUUID, arg) => {
   }
 }
 
+export var saveState = (state) => {
+  var yamlObj = yaml.dump(state);
+  fs.writeFileSync(dir + 'state.yml', yamlObj);
+
+}
+
+export var getState = () => {
+  if(!fs.existsSync(dir + 'state.yml')) {
+    return false;
+  }
+   var state = yaml.load(fs.readFileSync(dir + 'state.yml', 'utf8'));
+   return state;
+}
 //////////////////////Token Functions/////////////////////////
 export var checkToken = () => {
   var tempDir = isWin ? homedir + '\\AppData\\Local\\Temp' : '/tmp';
   var script = prefix + 'download 00007ccc-269b-4cd0-a0b1-6e5d700a8e5f -t ' + dir + 'token.txt -d ' + tempDir;
   if (!fs.existsSync(dir + 'token.txt')) {
-    return ({ consoleLog: '', tokenStatus: 'No Token File' });
+    return new Promise((resolve) => resolve({
+      consoleLog: '',
+      token: {
+        status: 'No Token File',
+        icon: <i className="fa fa-warning" aria-hidden="true" />,
+        colour: 'red'
+      }
+    }));
   }
   else {
     return new Promise((resolve, reject) => {
@@ -107,13 +197,35 @@ export var checkToken = () => {
         console.log(stderr);
         console.log('stdout: ' + stdout)
         if (stderr.includes('403 Client Error: FORBIDDEN')) {
-          resolve({ consoleLog: stderr, tokenStatus: 'Expired or invalid' });
+          resolve({
+            consoleLog: stderr,
+            token: {
+              status: 'Expired or invalid',
+              icon: <i className="fa fa-times-circle" aria-hidden="true" />,
+              colour: 'red'
+
+            }
+          });
         }
         else if (stdout.includes('Successfully downloaded')) {
-          resolve({ consoleLog: '', tokenStatus: 'Valid' });
+          resolve({
+            consoleLog: '',
+            token: {
+              status: 'Valid',
+              icon: <i className="fa fa-check" aria-hidden="true" />,
+              colour: 'green'
+            }
+          });
         }
         else {
-          resolve({ consoleLog: stderr, tokenStatus: 'Unknown' });
+          resolve({
+            consoleLog: stderr,
+            token: {
+              status: 'Unknown',
+              icon: <i className="fa fa-times-circle" aria-hidden="true" />,
+              colour: 'red'
+            }
+          });
         }
         if (error !== null) {
           reject('exec error: ' + error);
@@ -124,40 +236,106 @@ export var checkToken = () => {
 }
 
 export var saveToken = (tokenFile) => {
-  return new Promise(() => {
+  return new Promise((resolve, reject) => {
     fs.access(dir, () => {
       var readStream = fs.createReadStream(tokenFile);
-
       readStream.once('error', (err) => {
         console.log(err);
+        reject(err)
       });
 
       readStream.once('end', () => {
         console.log('done copying');
       });
       var tokenDest = dir + (isWin ? '\\token.txt' : 'token.txt');
-      readStream.pipe(fs.createWriteStream(tokenDest));
-      if (!isWin) {
-        var script = 'chmod 600 ' + dir + 'token.txt';
-        var cmd = exec(script, (error, stdout, stderr) => {
-          console.log(stderr);
-        });
-      }
-      else {
+      return new Promise((res) => res(readStream.pipe(fs.createWriteStream(tokenDest))))
+        .then(() => {
+          if (!isWin) {
+            var script = 'chmod 600 ' + dir + 'token.txt';
+            var cmd = exec(script, (error, stdout, stderr) => {
+              console.log(stderr);
+            });
+          }
+          else {
 
-      }
+          }
+          resolve();
+        });
     })
   });
 }
+//////////////////////Settings Functions//////////////////////
+export var saveSettings = (defaultSettings, state) => {
+  var obj = Object.keys(state).length === 0 ? defaultSettings : state;
+  var params = {
+    // connectionsParams: {
+    //   server: obj.server !== defaultSettings.server ? ' -s ' + obj.server : false,
+    //   port: obj.port !== defaultSettings.port ? ' -P ' + obj.port : false
+    // },
+    bothParams: {
+      numClientCons: obj.numClientCons <= 6 ? obj.numClientCons : 3,
+      // createLogFile: obj.createLogFile  ? ' --log-file ' + fixSpace(obj.logDestination) : false,
+      debugLogging: obj.debugLogging ? ' --debug' : false,
+      verboseLogging: obj.verboseLogging ? ' --v' : false
+    },
+    downloadParams: {
+      downloadDestination: ' -d ' + fixSpace(obj.downloadDestination),
+      blockSize: obj.blockSize !== defaultSettings.blockSize ?
+        ' --http-chunk-size ' + obj.blockSize : false,
 
+      saveInterval: obj.saveInterval !== defaultSettings.saveInterval ?
+        ' --save-interval ' + obj.saveInterval : false,
+
+      calcInSegAndCheckMd5: !obj.calcInSegAndCheckMd5 ? ' --no-segment-md5sums' : false,
+      checkMd5: !obj.checkMd5 ? ' --no-file-md5sum' : false,
+      // autoRetry: !obj.autoRetry ? ' --no-auto-retry' : false,
+      // numRetrys: obj.autoRetry ? ' --retry-amount ' + obj.numRetrys : false,
+      // retryInterval: obj.autoRetry ? ' --wait-time ' + obj.retryInterval : false
+    },
+    uploadParams: {
+      multipartUpload: !obj.multipartUpload ? ' --disable-multipart' : false,
+      partSize: obj.multipartUpload ? '-ps ' + partSize : false
+    }
+  }
+
+  var settings = obj;
+
+  var prefs = {
+    parameters: params,
+    settings: settings
+  };
+  var yamlObj = yaml.dump(prefs);
+  fs.writeFileSync(dir + 'prefs.yml', yamlObj);
+}
 //////////////////////Utility Functions///////////////////////
-export var killProcess = (processes) => {
-  processes.forEach(obj => {
-    clearInterval(obj.timer);
+export var getClientCons = () => {
+  var obj = yaml.load(fs.readFileSync(dir + 'prefs.yml', 'utf8'));
+  var cons = parseInt(obj.parameters.bothParams.numClientCons, 10);
+  return cons > 6 && cons < 1 ? 6 : cons;
+}
+
+export var isDirDefault = (type) => {
+  var obj = yaml.load(fs.readFileSync(dir + 'prefs.yml', 'utf8'));
+  if (type === 'download') {
+    return obj.parameters.downloadParams.downloadDestination === homedir;
+  }
+  else {
+    return obj.parameters.uploadParams.uploadSource === homedir;
+  }
+}
+export var killProcesses = (processes) => {
+  console.log(processes);
+  processes.forEach(process => killProcess(process))
+}
+
+export var killProcess = (process) => {
+  try {
+    process.callback();
+    clearInterval(process.timer);
     if (isWin) {
     }
     else {
-      exec('pkill -TERM -P ' + obj.process, { maxBuffer: 1024 * 1000 }, (error, stdout, stderr) => {
+      exec('pkill -TERM -P ' + process.pid, { maxBuffer: 1024 * 1000 }, (error, stdout, stderr) => {
         if (error !== null) {
           console.log('exec error: ' + error);
           console.log(stderr);
@@ -165,7 +343,7 @@ export var killProcess = (processes) => {
         }
       });
     }
-  })
+  } catch (e) { }
 }
 export var saveLog = (type, log) => {
   var prefs = yaml.load(fs.readFileSync(dir + 'prefs.yml', 'utf8'));
@@ -204,9 +382,8 @@ export var formatTime = (time) => {
 export var formatBytes = (bytes, decimals) => {
   if (bytes == 0) return '0 Bytes';
   var k = 1000,
-    dm = decimals + 1 || 3,
     sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'],
     i = Math.floor(Math.log(bytes) / Math.log(k));
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
 }
 
